refactor: vbridge `userWithdraw` hook allows multiple notes
